on-deck/requirements.txt
python-pptx

